# api/src/main/resources/rosetta-specifications-1.4.15/api.yaml
openapi: 3.0.2
info:
  version: 1.4.13
  title: Rosetta
  description: Build Once. Integrate Your Blockchain Everywhere.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /block/transaction:
    post:
      summary: Get a Block Transaction
      description: 'Get a transaction in a block by its Transaction Identifier. This endpoint should only be used when querying a node for a block does not return all transactions contained within it. All transactions returned by this endpoint must be appended to any transactions returned by the /block method by consumers of this data. Fetching a transaction by hash is considered an Explorer Method (which is classified under the Future Work section). This method can be used to let consumers to paginate results when the  block trasactions count is too big to be returned in a single BlockResponse. Calling this endpoint requires reference to a BlockIdentifier because transaction parsing can change depending on which block contains the transaction. For example, in Bitcoin it is necessary to know which block contains a transaction to determine the destination of fee payments. Without specifying a block identifier, the node would have to infer which block to use (which could change during a re-org). Implementations that require fetching previous transactions to populate the response (ex: Previous UTXOs in Bitcoin) may find it useful to run a cache within the Rosetta server in the /data directory (on a path that does not conflict with the node).'
      operationId: blockTransaction
      tags:
        - Block
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockTransactionRequest'
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockTransactionResponse'
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    NetworkIdentifier:
      description: The network_identifier specifies which network a particular object is associated with.
      type: object
      required:
        - blockchain
        - network
      properties:
        blockchain:
          type: string
          example: cardano
        network:
          description: If a blockchain has a specific chain-id or network identifier, it should go in this field. It is up to the client to determine which network-specific identifier is mainnet or testnet.
          type: string
          example: mainnet
        # sub_network_identifier:
        #   $ref: '#/components/schemas/SubNetworkIdentifier'
    SubNetworkIdentifier:
      description: In blockchains with sharded state, the SubNetworkIdentifier is required to query some object on a specific shard. This identifier is optional for all non-sharded blockchains.
      type: object
      required:
        - network
      properties:
        network:
          type: string
          example: shard 1
        metadata:
          type: object
          example:
            producer: '0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5'
    BlockIdentifier:
      description: The block_identifier uniquely identifies a block in a particular network.
      type: object
      required:
        - index
        - hash
      properties:
        index:
          description: This is also known as the block height.
          type: integer
          format: int64
          example: 1123941
        hash:
          type: string
          example: '0x1f2cc6c5027d2f201a5453ad1119574d2aed23a392654742ac3c78783c071f85'
          description: This should be normalized according to the case specified in the block_hash_case network options.
...
    TransactionIdentifier:
      description: The transaction_identifier uniquely identifies a transaction in a particular network and block or in the mempool.
      type: object
      required:
        - hash
      properties:
        hash:
          description: 'Any transactions that are attributable only to a block (ex: a block event) should use the hash of the block as the identifier.  This should be normalized according to the case specified in the transaction_hash_case in network options.'
          type: string
          example: '0x2f23fd8cca835af21f3ac375bac601f97ead75f2e79143bdf71fe2c4be043e8f'
    OperationIdentifier:
      description: The operation_identifier uniquely identifies an operation within a transaction.
      type: object
      required:
        - index
      properties:
        index:
          description: The operation index is used to ensure each operation has a unique identifier within a transaction. This index is only relative to the transaction and NOT GLOBAL. The operations in each transaction should start from index 0. To clarify, there may not be any notion of an operation index in the blockchain being described.
          type: integer
          format: int64
          minimum: 0
          example: 5
        network_index:
          description: Some blockchains specify an operation index that is essential for client use. For example, Bitcoin uses a network_index to identify which UTXO was used in a transaction. network_index should not be populated if there is no notion of an operation index in a blockchain (typically most account-based blockchains).
          type: integer
          format: int64
          minimum: 0
          example: 0
    AccountIdentifier:
      description: The account_identifier uniquely identifies an account within a network. All fields in the account_identifier are utilized to determine this uniqueness (including the metadata field, if populated).
      type: object
      required:
        - address
      properties:
        address:
          description: The address may be a cryptographic public key (or some encoding of it) or a provided username.
          type: string
          example: '0x3a065000ab4183c6bf581dc1e55a605455fc6d61'
        sub_account:
          $ref: '#/components/schemas/SubAccountIdentifier'
        metadata:
          $ref: '#/components/schemas/AccountIdentifierMetadata'
    SubAccountIdentifier:
      description: An account may have state specific to a contract address (ERC-20 token) and/or a stake (delegated balance). The sub_account_identifier should specify which state (if applicable) an account instantiation refers to.
      type: object
      required:
        - address
      properties:
        address:
          description: 'The SubAccount address may be a cryptographic value or some other identifier (ex: bonded) that uniquely specifies a SubAccount.'
          type: string
          example: '0x6b175474e89094c44da98b954eedeac495271d0f'
        metadata:
          $ref: '#/components/schemas/AccountIdentifierMetadata'
...
    Transaction:
      description: Transactions contain an array of Operations that are attributable to the same TransactionIdentifier.
      type: object
      required:
        - transaction_identifier
        - operations
      properties:
        transaction_identifier:
          $ref: '#/components/schemas/TransactionIdentifier'
        operations:
          type: array
          items:
            $ref: '#/components/schemas/Operation'
        related_transactions:
          type: array
          items:
            $ref: '#/components/schemas/RelatedTransaction'
        metadata:
          $ref: '#/components/schemas/TransactionMetadata'
    Operation:
      description: |
        Operations contain all balance-changing information within a transaction. They are always one-sided (only affect 1 AccountIdentifier) and can succeed or fail independently from a Transaction. Operations are used both to represent on-chain data (Data API) and to construct new transactions (Construction API), creating a standard interface for reading and writing to blockchains.\n        \n        For Cardano: Input operations use negative values for both ADA and native tokens (in tokenBundle metadata). Output operations use positive values. See https://cardano-foundation.github.io/cardano-rosetta-java/docs/user-guides/multi-assets for details.
      type: object
      required:
        - operation_identifier
        - type
      properties:
        operation_identifier:
          $ref: '#/components/schemas/OperationIdentifier'
        related_operations:
          description: Restrict referenced related_operations to identifier indices < the current operation_identifier.index. This ensures there exists a clear DAG-structure of relations. Since operations are one-sided, one could imagine relating operations in a single transfer or linking operations in a call tree.
          type: array
          items:
            $ref: '#/components/schemas/OperationIdentifier'
          example:
            - index: 1
            - index: 2
        type:
          description: Type is the network-specific type of the operation. Ensure that any type that can be returned here is also specified in the NetworkOptionsResponse. This can be very useful to downstream consumers that parse all block data.
          type: string
          example: Transfer
        status:
          description: Status is the network-specific status of the operation. Status is not defined on the transaction object because blockchains with smart contracts may have transactions that partially apply (some operations are successful and some are not). Blockchains with atomic transactions (all operations succeed or all operations fail) will have the same status for each operation. On-chain operations (operations retrieved in the `/block` and `/block/transaction` endpoints) MUST have a populated status field (anything on-chain must have succeeded or failed). However, operations provided during transaction construction (often times called "intent" in the documentation) MUST NOT have a populated status field (operations yet to be included on-chain have not yet succeeded or failed).
          type: string
          example: Reverted
        account:
          $ref: '#/components/schemas/AccountIdentifier'
        amount:
          $ref: '#/components/schemas/Amount'
        coin_change:
          $ref: '#/components/schemas/CoinChange'
        metadata:
          $ref: '#/components/schemas/OperationMetadata'
    Amount:
      description: Amount is some Value of a Currency. It is considered invalid to specify a Value without a Currency.
      type: object
      required:
        - value
        - currency
      properties:
        value:
          description: Value of the transaction in atomic units represented as an arbitrary-sized signed integer. For example, 1 ADA would be represented by a value of 1000000 (since ADA has 6 decimals).
          type: string
          example: '1238089899992'
        currency:
          $ref: '#/components/schemas/CurrencyResponse'
        metadata:
          type: object
...
    CurrencyResponse:
      description: Currency for response operations
      type: object
      required:
        - symbol
        - decimals
      properties:
        symbol:
          type: string
          example: ADA
        decimals:
          type: integer
          format: int32
          minimum: 0
          example: 8
        metadata:
          $ref: '#/components/schemas/CurrencyMetadataResponse'
...
    CurveType:
      description: 'CurveType is the type of cryptographic curve associated with a PublicKey. * secp256k1: SEC compressed - `33 bytes` (https://secg.org/sec1-v2.pdf#subsubsection.2.3.3) * secp256k1_bip340: x-only - `32 bytes`  (implicitly even `Y` coord. Secp256k1 compressed keys may be repurposed by dropping the first byte. (https://github.com/bitcoin/bips/blob/master/bip-0340.mediawiki#Public_Key_Generation)) * secp256r1: SEC compressed - `33 bytes` (https://secg.org/sec1-v2.pdf#subsubsection.2.3.3) * edwards25519: `y (255-bits) || x-sign-bit (1-bit)` - `32 bytes` (https://ed25519.cr.yp.to/ed25519-20110926.pdf) * tweedle: 1st pk : Fq.t (32 bytes) || 2nd pk : Fq.t (32 bytes) (https://github.com/CodaProtocol/coda/blob/develop/rfcs/0038-rosetta-construction-api.md#marshal-keys) * pallas: `x (255 bits) || y-parity-bit (1-bit) - 32 bytes` (https://github.com/zcash/pasta)'
      type: string
      enum:
        - secp256k1
        - secp256k1_bip340
        - secp256r1
        - edwards25519
        - tweedle
        - pallas
...
    CoinAction:
      description: CoinActions are different state changes that a Coin can undergo. When a Coin is created, it is coin_created. When a Coin is spent, it is coin_spent. It is assumed that a single Coin cannot be created or spent more than once.
      type: string
      enum:
        - coin_created
        - coin_spent
    CoinIdentifier:
      description: CoinIdentifier uniquely identifies a Coin.
      type: object
      required:
        - identifier
      properties:
        identifier:
          description: Identifier should be populated with a globally unique identifier of a Coin. In Bitcoin, this identifier would be transaction_hash:index.
          type: string
          example: 0x2f23fd8cca835af21f3ac375bac601f97ead75f2e79143bdf71fe2c4be043e8f:1
    CoinChange:
      description: CoinChange is used to represent a change in state of a some coin identified by a coin_identifier. This object is part of the Operation model and must be populated for UTXO-based blockchains. Coincidentally, this abstraction of UTXOs allows for supporting both account-based transfers and UTXO-based transfers on the same blockchain (when a transfer is account-based, don't populate this model).
      type: object
      required:
        - coin_identifier
        - coin_action
      properties:
        coin_identifier:
          $ref: '#/components/schemas/CoinIdentifier'
        coin_action:
          $ref: '#/components/schemas/CoinAction'
...
    RelatedTransaction:
      description: The related_transaction allows implementations to link together multiple transactions. An unpopulated network identifier indicates that the related transaction is on the same network.
      type: object
      required:
        - transaction_identifier
        - direction
      properties:
        network_identifier:
          $ref: '#/components/schemas/NetworkIdentifier'
        transaction_identifier:
          $ref: '#/components/schemas/TransactionIdentifier'
        direction:
          $ref: '#/components/schemas/Direction'
    Direction:
      description: Used by RelatedTransaction to indicate the direction of the relation (i.e. cross-shard/cross-network sends may reference `backward` to an earlier transaction and async execution may reference `forward`). Can be used to indicate if a transaction relation is from child to parent or the reverse.
      type: string
      enum:
        - forward
        - backward
...
    BlockTransactionRequest:
      description: A BlockTransactionRequest is used to fetch a Transaction included in a block that is not returned in a BlockResponse.
      type: object
      required:
        - network_identifier
        - block_identifier
        - transaction_identifier
      properties:
        network_identifier:
          $ref: '#/components/schemas/NetworkIdentifier'
        block_identifier:
          $ref: '#/components/schemas/BlockIdentifier'
        transaction_identifier:
          $ref: '#/components/schemas/TransactionIdentifier'
    BlockTransactionResponse:
      description: A BlockTransactionResponse contains information about a block transaction.
      type: object
      required:
        - transaction
      properties:
        transaction:
          $ref: '#/components/schemas/Transaction'
...
    Error:
      description: Instead of utilizing HTTP status codes to describe node errors (which often do not have a good analog), rich errors are returned using this object. Both the code and message fields can be individually used to correctly identify an error. Implementations MUST use unique values for both fields.
      type: object
      required:
        - code
        - message
        - retriable
      properties:
        code:
          description: Code is a network-specific error code. If desired, this code can be equivalent to an HTTP status code.
          type: integer
          format: int32
          minimum: 0
          example: 12
        message:
          description: Message is a network-specific error message. The message MUST NOT change for a given code. In particular, this means that any contextual information should be included in the details field.
          type: string
          example: Invalid account format
        description:
          description: Description allows the implementer to optionally provide additional information about an error. In many cases, the content of this field will be a copy-and-paste from existing developer documentation. Description can ONLY be populated with generic information about a particular type of error. It MUST NOT be populated with information about a particular instantiation of an error (use `details` for this). Whereas the content of Error.Message should stay stable across releases, the content of Error.Description will likely change across releases (as implementers improve error documentation). For this reason, the content in this field is not part of any type assertion (unlike Error.Message).
          type: string
          example: This error is returned when the requested AccountIdentifier is improperly formatted.
        retriable:
          description: An error is retriable if the same request may succeed if submitted again.
          type: boolean
        details:
          type: object
          description: Often times it is useful to return context specific to the request that caused the error (i.e. a sample of the stack trace or impacted account) in addition to the standard error message.
          example:
            address: '0x1dcc4de8dec75d7aab85b567b6'
            error: not base64
...
    TransactionMetadata:
      description: Transactions that are related to other transactions (like a cross-shard transaction) should include the tranaction_identifier of these transactions in the metadata.
      type: object
      properties:
        size:
          description: Size is the size of the transaction in bytes.
          type: integer
          format: int64
        scriptSize:
          description: ScriptSize is the size of the transaction script in bytes.
          type: integer
          format: int64
      example:
        size: 12378
        lockTime: 1582272577
    OperationMetadata:
      description: |
        OperationMetadata provides additional data needed for an Operation. For Cardano-specific operations, this includes native token bundles, staking credentials, and governance parameters. 
        See https://cardano-foundation.github.io/cardano-rosetta-java/docs/core-concepts/cardano-addons for Cardano-specific metadata documentation.
      type: object
      properties:
        withdrawalAmount:
          description: WithdrawalAmount is the amount of funds withdrawn from an account. This field should only be populated if the operation is a withdrawal.
          $ref: '#/components/schemas/Amount'
        depositAmount:
          description: DepositAmount is the amount of funds deposited into an account. This field should only be populated if the operation is a deposit.
          $ref: '#/components/schemas/Amount'
        refundAmount:
          description: RefundAmount is the amount of funds refunded to an account. This field should only be populated if the operation is a refund.
          $ref: '#/components/schemas/Amount'
        staking_credential:
          $ref: '#/components/schemas/PublicKey'
        pool_key_hash:
          description: Pool key as hex bytes.
          type: string
        drep:
          description: DRep parameters for vote delegation operations
          $ref: '#/components/schemas/DRepParams'
        poolGovernanceVoteParams:
          description: PoolGovernanceVoteParams contains parameters related to casting a governance vote by a pool operator.
          type: object
          $ref: '#/components/schemas/PoolGovernanceVoteParams'
        epoch:
          description: Epoch
          type: integer
        tokenBundle:
          description: |
            Native token bundle associated with the operation. Token values MUST be negative for input operations (spending) and positive for output operations (receiving). 
            See documentation at https://cardano-foundation.github.io/cardano-rosetta-java/docs/user-guides/multi-assets for details.
          type: array
          items:
            $ref: '#/components/schemas/TokenBundleItem'
        poolRegistrationCert:
          description: PoolRegistrationCert
          type: string
        poolRegistrationParams:
          $ref: '#/components/schemas/PoolRegistrationParams'
        voteRegistrationMetadata:
          $ref: '#/components/schemas/VoteRegistrationMetadata'
    PoolGovernanceVoteParams:
      description: Parameters for a governance vote by a pool operator.
      type: object
      properties:
        governance_action_hash:
          description: GovAction is a unique identifier for governance action
          $ref: '#/components/schemas/GovActionParams'
        pool_credential:
          description: Pool Credential is the credential of the pool operator casting the vote.
          $ref: '#/components/schemas/PublicKey'
        vote:
          description: GovVoteParams is the vote cast by the pool operator, which can be "yes", "no", or "abstain".
          $ref: '#/components/schemas/GovVoteParams'
        vote_rationale:
          description: GovVoteRationale is an optional rationale for the vote cast by the pool operator.
          $ref: '#/components/schemas/GovVoteRationaleParams'
    GovActionParams:
      description: Unique identifier for a governance action, consisting of tx_id and index concatenated.
      type: string
      example: "df58f714c0765f3489afb6909384a16c31d600695be7e86ff9c59cf2e8a48c7900"
    GovVoteParams:
      description: The vote cast by the voter.
      type: string
      enum:
        - "yes"
        - "no"
        - "abstain"
    GovVoteRationaleParams:
      description: The rationale for the vote cast by the voter.
      type: object
      properties:
        url:
          type: string
          description: URL of the anchor.
        data_hash:
          type: string
          description: Hash of the anchor data.
    VoteRegistrationMetadata:
      description: VoteRegistrationMetadata
      type: object
      properties:
        stakeKey:
          description: VotePublicKey
          $ref: '#/components/schemas/PublicKey'
        votingkey:
          description: votingkey
          $ref: '#/components/schemas/PublicKey'
        rewardAddress:
          description: rewardAddress
          type: string
        votingNonce:
          description: votingNonce
          type: integer
        votingSignature:
          description: votingSignature
          type: string
    TokenBundleItem:
      description: |
        Represents a collection of native tokens under a single policy ID. Used in tokenBundle metadata to track native assets in operations.
        Note: Minting and burning are not supported through Construction API - only transfers of existing tokens.
      type: object
      properties:
        policyId:
          description: The policy ID that controls these native tokens (hex string)
          type: string
        tokens:
          description: |
            Array of token amounts under this policy. Values must be negative for inputs (spending) and positive for outputs (receiving).
            Symbol should be the ASCII representation of the token name.
          type: array
          items:
            $ref: '#/components/schemas/Amount'
    DRepTypeParams:
      description: Type of DRep delegation (key_hash, script_hash, no_confidence, or abstain)
      type: string
      enum:
        - key_hash
        - script_hash
        - no_confidence
        - abstain
    DRepParams:
        description: |
          Delegated Representative (DRep) parameters for vote delegation.
          Supports CIP-129 format: accepts both 29-byte hashes (with prefix) and 28-byte raw hashes.
          The type field is required for all formats.
          See https://cardano-foundation.github.io/cardano-rosetta-java/docs/user-guides/drep-delegation for details.
        type: object
        required:
          - type
        properties:
            id:
              description: |
                DRep identifier. For key_hash and script_hash types:
                - Accepts 29-byte hash with CIP-129 prefix (e.g., "03" for key_hash)
                - Accepts 28-byte raw hash without prefix
                - The API automatically strips the prefix if present
                Not required for abstain and no_confidence types.
              type: string
              example: "74984fae4ca1715fa1f8759f9d871015ac87f449a85dea6cf9956da1"
            type:
              description: Type of DRep delegation. Required for all operations.
              $ref: '#/components/schemas/DRepTypeParams'
    PoolRegistrationParams:
      description: PoolRegistrationParams
      type: object
      properties:
        vrfKeyHash:
          description: VrfKeyHash
          type: string
        rewardAddress:
          description: RewardAddress
          type: string
        pledge:
          description: Pledge
          type: string
        cost:
          description: Cost
          type: string
        poolOwners:
          description: PoolOwners
          type: array
          items:
            type: string
        relays:
          description: Relays
          type: array
          items:
            $ref: '#/components/schemas/Relay'
        margin:
          description: Margin
          $ref: '#/components/schemas/PoolMargin'
        marginPercentage:
          description: MarginPercentage
          type: string
        poolMetadata:
          description: PoolMetadata
          $ref: '#/components/schemas/PoolMetadata'
    PoolMetadata:
      description: PoolMetadata
      type: object
      properties:
        url:
          description: url
          type: string
        hash:
          description: hash
          type: string
    Relay:
      description: Relay
      type: object
      properties:
        ipv4:
          description: Ipv4
          type: string
        ipv6:
          description: Ipv6
          type: string
        dnsName:
          description: DnsName
          type: string
        port:
          description: Port
          type: integer
        type:
          description: Type
          type: string
    PoolMargin:
      description: PoolMargin
      type: object
      properties:
        numerator:
          description: Numerator
          type: string
        denominator:
          description: Denominator
          type: string
...
    PublicKey:
      description: StakingCredential / PoolCredential or a PublicKey.
      type: object
      properties:
        hex_bytes:
          type: string
          description: Hex-encoded staking key hash.
        curve_type:
          $ref: '#/components/schemas/CurveType'
...
    AccountIdentifierMetadata:
      description: AccountIdentifierMetadata
      type: object
      properties:
        chain_code:
          description: 'ChainCode'
          type: string
...
    LogoType:
      description: 'Token logo representation that can be either base64 encoded data (CIP_26) or URL (CIP_68)'
      type: object
      required:
        - format
        - value
      properties:
        format:
          description: 'The format of the logo data'
          type: string
          enum:
            - base64
            - url
        value:
          description: 'The logo content - either base64 encoded string or URL'
          type: string
    CurrencyMetadataResponse:
      description: 'Cardano-specific response metadata for native tokens'
      type: object
      required:
        - policyId
      properties:
        policyId:
          description: 'The policy ID that controls this native token (hex string). Required for all non-ADA currencies.'
          type: string
        subject:
          description: 'The base16-encoded policyId + base16-encoded assetName'
          type: string
        name:
          description: 'Token Registry data representing an asset name'
          type: string
        description:
          description: 'Token Registry data representing an asset description'
          type: string
        ticker:
          description: 'Token Registry data representing an asset ticker symbol'
          type: string
        url:
          description: 'Token Registry data representing an asset''s project url'
          type: string
        logo:
          description: 'Token Registry data representing an asset''s logo'
          $ref: '#/components/schemas/LogoType'
        version:
          description: 'Token Registry data representing an asset''s version'
          type: number
...